Revert "refactor: Extract Business Logic from Components into Dedicated Hooks…"
This reverts commit 1bad275c145501143cd1ad1a477aa8000b1d3884.
