+++ components/HrmConnectionPanel.tsx
@@ -1,6 +1,6 @@
-// components/HrmConnectionPanel.tsx
+// File: app/components/dashboard/HrmConnectionPanel.tsx
 'use client'
-import { useCallback } from 'react'
+import { useMemo, useEffect, useCallback, useRef } from 'react'
 import Box from '@mui/material/Box'
 import Skeleton from '@mui/material/Skeleton'
 import Typography from '@mui/material/Typography'
@@ -11,7 +11,6 @@ import IconButton from '@mui/material/IconButton'
 import SettingsIcon from '@mui/icons-material/Settings'
 import useBluetoothHRM from '@/hooks/useBluetoothHRM'
 import { useWebSocket } from '@/context/WebSocketContext'
-import { useHrmPanel } from '@/hooks/useHrmPanel'
 import { CONNECT_HR_MONITOR_TITLE } from '@/utils/constants'
 import ConnectHRMonitorButton from './ConnectHRMonitorButton'
 import HRMonitorStatusIndicator from './HRMonitorStatusIndicator'
@@ -21,7 +20,7 @@ const HrmConnectionPanel = () => {
   const { data: session } = useSession()
   const [userSettings] = useUserSettings()
   const { hrmData, connectionStatus, activeAlerts } = useWebSocket()
-
+  const autoConnectAttempted = useRef(false)
   const {
     connectAndStream,
     disconnect,
@@ -34,25 +33,64 @@ const HrmConnectionPanel = () => {
     userAge: userSettings.userAge || 30,
   })
 
-  const { tileData, isLoading } = useHrmPanel({
-    hrmData,
-    activeAlerts,
-    connectionStatus,
-    deviceStatus,
-    connectAndStream,
-    userSettings,
-  })
-
   const handleConnect = useCallback(() => {
     const userName =
       session?.user?.name || userSettings.userName || 'Unknown User'
     const userAge = userSettings.userAge || 30
     connectAndStream(userName, userAge).catch((error) => {
+      // It's common for the requestDevice promise to be cancelled by the user.
+      // We catch it here to prevent an unhandled rejection error in the console.
       if (error.name !== 'NotFoundError') {
         console.error('Failed to connect to HRM device:', error)
       }
     })
   }, [session, userSettings, connectAndStream])
+
+  useEffect(() => {
+    // Auto-connect logic: Use `getDevices()` for gesture-less reconnection.
+    const autoConnect = async () => {
+      if (
+        connectionStatus === 'Connected' &&
+        deviceStatus === 'Disconnected' &&
+        !autoConnectAttempted.current
+      ) {
+        autoConnectAttempted.current = true
+        const userName =
+          session?.user?.name || userSettings.userName || 'Unknown User'
+        const userAge = userSettings.userAge || 30
+        await connectAndStream(userName, userAge)
+      }
+    }
+
+    autoConnect()
+  }, [connectionStatus, deviceStatus, connectAndStream, session, userSettings])
+
+  const tileData = useMemo(() => {
+    // Filter out users with placeholder names or no identity
+    return hrmData
+      .filter((user) => {
+        const isPlaceholderName = !!user.name && /new user/i.test(user.name)
+        const hasNoIdentity = user.name == null
+        return !(isPlaceholderName || hasNoIdentity)
+      })
+      .map((user) => {
+        const matchingAlert = activeAlerts.find(
+          (alert) =>
+            alert.clientId === user.clientId &&
+            (alert.code === 'BAD_PLACEMENT' || alert.code === 'HRM_STALE')
+        )
+
+        return {
+          ...user,
+          isAlerting: !!matchingAlert,
+          alertMessage: matchingAlert?.message,
+        }
+      })
+  }, [hrmData, activeAlerts])
+
+  const isLoading =
+    connectionStatus === 'Connecting...' ||
+    connectionStatus === 'Reconnecting...'
 
   return (
     <Box
@@ -70,7 +108,7 @@ const HrmConnectionPanel = () => {
               display: 'flex',
               flexDirection: 'column',
               width: { xs: '100%', sm: 'calc(50% - 8px)' },
-              height: '100%',
+              height: '100%', // Ensure the container fills the grid cell
               gap: 2,
               p: 2,
               border: 1,
@@ -125,7 +163,7 @@ const HrmConnectionPanel = () => {
             sx={{
               width: {
                 xs: '100%',
-                sm: 'calc(50% - 8px)',
+                sm: 'calc(50% - 8px)', // Adjusted for 16px gap (gap: 2)
               },
             }}
           >
